+++ lib/keymap.vi.sh
@@ -902,7 +902,7 @@ function ble/widget/vi-command/edit-and-execute-command {
   ble/keymap:vi/repeat/clear-insert
   [[ $_ble_keymap_vi_reg_record ]] &&
     ble/widget/vi_nmap/record-register
-  ble/widget/edit-and-execute-command
+  ble/widget/edit-and-execute-command vi
 }
 
 #------------------------------------------------------------------------------