+++ docs/ChangeLog.md
@@ -72,6 +72,7 @@
 - edit: support `bleopt history_default_point={preserve,begin,end,near,far,{beginning,end}-of-line,preserve-column,...}` (requested by miltieIV2) `#D2297` 37291ff1
 - edit: support `bleopt undo_point={first,last,near,auto}` `#D2303` xxxxxxxx
 - keymap/vi: add readline-compatible widgets for `vi_imap` and `vi_nmap` (requested by excited-bore) `#D2304` xxxxxxxx
+- edit: support bash-5.2 readline bindable function `vi-edit-and-execute-command` `#D2306` xxxxxxxx
 
 ## Changes